Add HACS support and icon (#34)
* Update vacuum.py with Battery Charge Level sensor

add sensor for battery charge level so it can be used on the main Entities Card beside the current vac status so you do not have to click the entitie to find out the charge level now that off peak charging has been added to the main app.

* Update hacs.json

try to add an icon for the project

* Add files via upload

add the actual icon this commit. oops

* Update vacuum.py

fix deprecation warning in Detected that custom integration 'robovac_mqtt' is setting state directly

* Update vacuum.py

1st part of fix for  AttributeError: 'int' object has no attribute 'lower'

* Update MqttConnect.py

2nd part fix for AttributeError: 'int' object has no attribute 'lower'

* Update SharedConnect.py

3rd and final part of AttributeError: 'int' object has no attribute 'lower'

* assets/

* Update README.md

* Delete eufy-battery.png

wrong place. should have been in folder.

* Add files via upload

* Update hacs.json

* Delete icon.png

---------

Co-authored-by: Tobias Jeppesen <github@tobiasjeppesen.com>